This commit achieves a number of objectives.
First and foremost, it introduces full 1.12 support for RecipeManager
recipe control. It achieves this through the introduction of two NMS
support classes, one to handle recipe matching for 1.12, and another to
support recipe iteration and removal for 1.12.

In Minecraft 1.12, due to advancements, Recipes have been significantly
reworked. The exact ordering of recipes is now important; specifically,
the vanilla recipe exact ordering is critical. So, the new 1.12 support
throughout RecipeManager, and especially the NMS sections, allow for
recipe "removal" simply through suppression. The recipes are modified
in-place with versions that will never match during crafting; ordering,
however, is preserved so Client-side advancements are less harmed by the
transition.

Note however, that if you do remove a recipe, clientside advancements
may have some artifacts involved when the player attempts to "pick" that
recipe from their recipe books once unlocked.

Additionally, if you replace a recipe, advancements for that recipe will
not work in the future; however, impairment will be restricted to those
recipes alone.

Future efforts might be necessary to better preserve advancements; for
instance, replace could be done "in-place" using an expansion of the
techniques used in RecipeIteratorV1_12. Removals could be similarly
hooked, suppressed by RecipeManager and otherwise untouched in the list.

I'll be exploring those options next.

Pending that, future efforts may be required to add new methods to hook
into advancements and inform the client of recipe removal. For now, this
seeks to split the difference, and allow largely backwards compatible
recipe manipulation including full support for Minecraft 1.12.

Thanks to Haveric for brainstorming and continued support of
RecipeManager and md_5 for the comment pointing to
necessity of preserving total ordering.