Fixes (most) 1.12 support issues (#70)
* This commit achieves a number of objectives.

First and foremost, it introduces full 1.12 support for RecipeManager
recipe control. It achieves this through the introduction of two NMS
support classes, one to handle recipe matching for 1.12, and another to
support recipe iteration and removal for 1.12.

In Minecraft 1.12, due to advancements, Recipes have been significantly
reworked. The exact ordering of recipes is now important; specifically,
the vanilla recipe exact ordering is critical. So, the new 1.12 support
throughout RecipeManager, and especially the NMS sections, allow for
recipe "removal" simply through suppression. The recipes are modified
in-place with versions that will never match during crafting; ordering,
however, is preserved so Client-side advancements are less harmed by the
transition.

Note however, that if you do remove a recipe, clientside advancements
may have some artifacts involved when the player attempts to "pick" that
recipe from their recipe books once unlocked.

Additionally, if you replace a recipe, advancements for that recipe will
not work in the future; however, impairment will be restricted to those
recipes alone.

Future efforts might be necessary to better preserve advancements; for
instance, replace could be done "in-place" using an expansion of the
techniques used in RecipeIteratorV1_12. Removals could be similarly
hooked, suppressed by RecipeManager and otherwise untouched in the list.

I'll be exploring those options next.

Pending that, future efforts may be required to add new methods to hook
into advancements and inform the client of recipe removal. For now, this
seeks to split the difference, and allow largely backwards compatible
recipe manipulation including full support for Minecraft 1.12.

Thanks to Haveric for brainstorming and continued support of
RecipeManager and md_5 for the comment pointing to
necessity of preserving total ordering.

* Fixed reloading, appears to work consistently now with no loss of fidelity of Vanilla recipes. Also improved spoofing of invalid recipes for removing existing in-place.

* Discovered simple workaround for loss of recipe fidelity. While not
future-perfect, the simplest route is to simply give you recipes defined
names, and avoid the randomly generated name that would otherwise be
assigned. In this way, connectivity is maintained.

* This removes the try-catches around Bukkit calls; it also adds in-place
replacement, with no removal, of vanilla recipes. In this way, there is
absolutely no disruption of Advancements even if you remove or replace a
vanilla recipe in 1.12.

I also modified how rmreload works, it _should_ preserve other recipes
now although ordering is not guaranteed -- as always, though, MC recipe
ordering will be preserved during reload.

Still to do is adding a warning that rereload might be unstable for
1.12.

* Adding Coversation for reload in 1.12, as discussed.